Move testing/ui and testing/store to private/node
