+++ packages/cli-kit/src/private/node/ui/components/ConcurrentOutput.test.tsx
@@ -1,6 +1,6 @@
 import ConcurrentOutput from './ConcurrentOutput.js'
 import {unstyled} from '../../../../output.js'
-import {getLastFrameAfterUnmount} from '../../../../testing/ui.js'
+import {getLastFrameAfterUnmount} from '../../testing/ui.js'
 import {AbortController, AbortSignal} from '../../../../public/node/abort.js'
 import React from 'react'
 import {describe, expect, test} from 'vitest'
