+++ packages/cli-kit/src/private/node/ui/components/SelectInput.test.tsx
@@ -1,5 +1,5 @@
 import SelectInput from './SelectInput.js'
-import {sendInputAndWait, sendInputAndWaitForChange, waitForInputsToBeReady} from '../../../../testing/ui.js'
+import {sendInputAndWait, sendInputAndWaitForChange, waitForInputsToBeReady} from '../../testing/ui.js'
 import {describe, expect, test, vi} from 'vitest'
 import React from 'react'
 import {render} from 'ink-testing-library'
