+++ packages/cli-kit/src/private/node/ui/components/SelectPrompt.test.tsx
@@ -1,5 +1,5 @@
 import {SelectPrompt} from './SelectPrompt.js'
-import {getLastFrameAfterUnmount, sendInputAndWaitForChange, waitForInputsToBeReady} from '../../../../testing/ui.js'
+import {getLastFrameAfterUnmount, sendInputAndWaitForChange, waitForInputsToBeReady} from '../../testing/ui.js'
 import {unstyled} from '../../../../output.js'
 import {describe, expect, test, vi} from 'vitest'
 import React from 'react'
